+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fix parsing of import declarations in module headers. ([#136](https://github.com/fabianschuiki/moore/issues/136))
 
 ## 0.7.0 - 2020-04-30
 ### Added